+++ WeaponAction.cs
@@ -1,0 +1,222 @@
+﻿using CounterStrikeSharp.API;
+using CounterStrikeSharp.API.Core;
+using CounterStrikeSharp.API.Modules.Memory;
+using CounterStrikeSharp.API.Modules.Utils;
+
+namespace WeaponPaints
+{
+	public partial class WeaponPaints
+	{
+		internal static void ChangeWeaponAttributes(CBasePlayerWeapon? weapon, CCSPlayerController? player, bool isKnife = false)
+		{
+			if (weapon == null || !weapon.IsValid || !Utility.IsPlayerValid(player)) return;
+
+			int playerIndex = (int)player!.EntityIndex!.Value.Value;
+
+			if (!gPlayerWeaponsInfo.ContainsKey(playerIndex)) return;
+
+			if (isKnife && !g_playersKnife.ContainsKey(playerIndex) || isKnife && g_playersKnife[playerIndex] == "weapon_knife") return;
+
+			int weaponDefIndex = weapon.AttributeManager.Item.ItemDefinitionIndex;
+
+			if (_config.Additional.GiveRandomSkin &&
+				 !gPlayerWeaponsInfo[playerIndex].ContainsKey(weaponDefIndex))
+			{
+				// Random skins
+				weapon.AttributeManager.Item.ItemID = 16384;
+				weapon.AttributeManager.Item.ItemIDLow = 16384 & 0xFFFFFFFF;
+				weapon.AttributeManager.Item.ItemIDHigh = weapon.AttributeManager.Item.ItemIDLow >> 32;
+				weapon.FallbackPaintKit = GetRandomPaint(weaponDefIndex);
+				weapon.FallbackSeed = 0;
+				weapon.FallbackWear = 0.0f;
+				if (!isKnife && weapon.CBodyComponent != null && weapon.CBodyComponent.SceneNode != null)
+				{
+					var skeleton = GetSkeletonInstance(weapon.CBodyComponent.SceneNode);
+					skeleton.ModelState.MeshGroupMask = 2;
+				}
+				return;
+			}
+
+			if (!gPlayerWeaponsInfo[playerIndex].ContainsKey(weaponDefIndex)) return;
+			WeaponInfo weaponInfo = gPlayerWeaponsInfo[playerIndex][weaponDefIndex];
+			//Log($"Apply on {weapon.DesignerName}({weapon.AttributeManager.Item.ItemDefinitionIndex}) paint {gPlayerWeaponPaints[steamId.SteamId64][weapon.AttributeManager.Item.ItemDefinitionIndex]} seed {gPlayerWeaponSeed[steamId.SteamId64][weapon.AttributeManager.Item.ItemDefinitionIndex]} wear {gPlayerWeaponWear[steamId.SteamId64][weapon.AttributeManager.Item.ItemDefinitionIndex]}");
+			weapon.AttributeManager.Item.ItemID = 16384;
+			weapon.AttributeManager.Item.ItemIDLow = 16384 & 0xFFFFFFFF;
+			weapon.AttributeManager.Item.ItemIDHigh = weapon.AttributeManager.Item.ItemIDLow >> 32;
+			weapon.FallbackPaintKit = weaponInfo.Paint;
+			weapon.FallbackSeed = weaponInfo.Seed;
+			weapon.FallbackWear = weaponInfo.Wear;
+			if (!isKnife && weapon.CBodyComponent != null && weapon.CBodyComponent.SceneNode != null)
+			{
+				var skeleton = GetSkeletonInstance(weapon.CBodyComponent.SceneNode);
+				skeleton.ModelState.MeshGroupMask = 2;
+			}
+		}
+		internal static void GiveKnifeToPlayer(CCSPlayerController? player)
+		{
+			if (!_config.Additional.KnifeEnabled || player == null || !player.EntityIndex.HasValue || !player.IsValid) return;
+			if (g_playersKnife.TryGetValue((int)player.EntityIndex.Value.Value, out var knife))
+			{
+				player.GiveNamedItem(knife);
+			}
+			else if (_config.Additional.GiveRandomKnife)
+			{
+				var knifeTypes = weaponList.Where(pair => pair.Key.StartsWith("weapon_knife") || pair.Key.StartsWith("weapon_bayonet")).ToDictionary(pair => pair.Key, pair => pair.Value);
+
+				Random random = new();
+				int index = random.Next(knifeTypes.Count);
+				var randomKnifeClass = knifeTypes.Keys.ElementAt(index);
+
+				player.GiveNamedItem(randomKnifeClass);
+			}
+			else
+			{
+				var defaultKnife = (CsTeam)player.TeamNum == CsTeam.Terrorist ? "weapon_knife_t" : "weapon_knife";
+				player.GiveNamedItem(defaultKnife);
+			}
+		}
+		internal void RemovePlayerKnife(CCSPlayerController? player)
+		{
+			if (player == null || !player.IsValid || !player.PawnIsAlive) return;
+			if (player.PlayerPawn.Value.WeaponServices == null || player.PlayerPawn.Value.ItemServices == null) return;
+
+			var weapons = player.PlayerPawn.Value.WeaponServices.MyWeapons;
+			if (weapons != null && weapons.Count > 0)
+			{
+				CCSPlayer_ItemServices service = new CCSPlayer_ItemServices(player.PlayerPawn.Value.ItemServices.Handle);
+				//var dropWeapon = VirtualFunction.CreateVoid<nint, nint>(service.Handle, GameData.GetOffset("CCSPlayer_ItemServices_DropActivePlayerWeapon"));
+
+				foreach (var weapon in weapons)
+				{
+					if (weapon != null && weapon.IsValid && weapon.Value.IsValid)
+					{
+						//if (weapon.Value.AttributeManager.Item.ItemDefinitionIndex == 42 || weapon.Value.AttributeManager.Item.ItemDefinitionIndex == 59)
+						if (weapon.Value.DesignerName.Contains("knife") || weapon.Value.DesignerName.Contains("bayonet"))
+						{
+							if (!weapon.Value.EntityIndex.HasValue) return;
+							int weaponEntityIndex = (int)weapon.Value.EntityIndex!.Value.Value;
+							NativeAPI.IssueClientCommand((int)player.EntityIndex!.Value.Value - 1, "slot3");
+							AddTimer(0.35f, () => service.DropActivePlayerWeapon(weapon.Value));
+
+							AddTimer(1.0f, () =>
+							{
+								CEntityInstance? knife = Utilities.GetEntityFromIndex<CEntityInstance>(weaponEntityIndex);
+								if (knife != null && knife.IsValid)
+								{
+									knife.Remove();
+								}
+							});
+
+							break;
+						}
+					}
+				}
+			}
+		}
+		internal void RefreshPlayerKnife(CCSPlayerController? player)
+		{
+			if (player == null || !player.IsValid || !player.PawnIsAlive) return;
+			if (player.PlayerPawn.Value.WeaponServices == null || player.PlayerPawn.Value.ItemServices == null) return;
+
+			var weapons = player.PlayerPawn.Value.WeaponServices.MyWeapons;
+			if (weapons != null && weapons.Count > 0)
+			{
+				CCSPlayer_ItemServices service = new CCSPlayer_ItemServices(player.PlayerPawn.Value.ItemServices.Handle);
+				//var dropWeapon = VirtualFunction.CreateVoid<nint, nint>(service.Handle, GameData.GetOffset("CCSPlayer_ItemServices_DropActivePlayerWeapon"));
+
+				foreach (var weapon in weapons)
+				{
+					if (weapon != null && weapon.IsValid && weapon.Value.IsValid)
+					{
+						//if (weapon.Value.AttributeManager.Item.ItemDefinitionIndex == 42 || weapon.Value.AttributeManager.Item.ItemDefinitionIndex == 59)
+						if (weapon.Value.DesignerName.Contains("knife") || weapon.Value.DesignerName.Contains("bayonet"))
+						{
+							if (!weapon.Value.EntityIndex.HasValue) return;
+							int weaponEntityIndex = (int)weapon.Value.EntityIndex!.Value.Value;
+							NativeAPI.IssueClientCommand((int)player.EntityIndex!.Value.Value - 1, "slot3");
+							AddTimer(0.35f, () => service.DropActivePlayerWeapon(weapon.Value));
+
+							AddTimer(1.0f, () =>
+							{
+								CEntityInstance? knife = Utilities.GetEntityFromIndex<CEntityInstance>(weaponEntityIndex);
+								if (knife != null && knife.IsValid)
+								{
+									knife.Remove();
+
+									if (player.PawnIsAlive)
+										GiveKnifeToPlayer(player);
+								}
+							});
+
+							break;
+						}
+					}
+				}
+			}
+		}
+
+		internal void RefreshSkins(CCSPlayerController? player)
+		{
+			if (!Utility.IsPlayerValid(player) || !player!.PawnIsAlive) return;
+
+			AddTimer(0.18f, () => NativeAPI.IssueClientCommand((int)player.EntityIndex!.Value.Value - 1, "slot3"));
+			AddTimer(0.25f, () => NativeAPI.IssueClientCommand((int)player.EntityIndex!.Value.Value - 1, "slot2"));
+			AddTimer(0.38f, () => NativeAPI.IssueClientCommand((int)player.EntityIndex!.Value.Value - 1, "slot1"));
+		}
+		internal static bool PlayerHasKnife(CCSPlayerController? player)
+		{
+			if (!WeaponPaints._config.Additional.KnifeEnabled) return false;
+
+			if (player == null || !player.IsValid)
+			{
+				return false;
+			}
+
+			var weapons = player.PlayerPawn.Value.WeaponServices!.MyWeapons;
+			if (weapons == null || weapons.Count <= 0) return false;
+			foreach (var weapon in weapons)
+			{
+				if (weapon != null && weapon.IsValid && weapon.Value.IsValid)
+				{
+					if (weapon.Value.DesignerName.Contains("knife") || weapon.Value.DesignerName.Contains("bayonet"))
+					{
+						return true;
+					}
+				}
+			}
+			return false;
+		}
+
+		private static CSkeletonInstance GetSkeletonInstance(CGameSceneNode node)
+		{
+			Func<nint, nint> GetSkeletonInstance = VirtualFunction.Create<nint, nint>(node.Handle, 8);
+			return new CSkeletonInstance(GetSkeletonInstance(node.Handle));
+		}
+
+		private static int GetRandomPaint(int defindex)
+		{
+			Random rnd = new Random();
+
+			if (WeaponPaints.skinsList != null)
+			{
+				// Filter weapons by the provided defindex
+				var filteredWeapons = WeaponPaints.skinsList.FindAll(w => w["weapon_defindex"]?.ToString() == defindex.ToString());
+
+				if (filteredWeapons.Count > 0)
+				{
+					var randomWeapon = filteredWeapons[rnd.Next(filteredWeapons.Count)];
+					if (int.TryParse(randomWeapon["paint"]?.ToString(), out int paintValue))
+					{
+						return paintValue;
+					}
+					else
+					{
+						return 0;
+					}
+
+				}
+			}
+			return 0;
+		}
+	}
+}
