+++ WeaponInfo.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WeaponPaints
+{
+	public class WeaponInfo
+	{
+		public int Paint { get; set; }
+		public int Seed { get; set; }
+		public float Wear { get; set; }
+	}
+}
